+++ django_backend/game/templates/game/index.html
@@ -46,7 +46,6 @@
             });
         </script>
         
-        
         {% if user.is_authenticated %}
         <a href="#" onClick="document.getElementById('logoutform').submit()">
             <button type="button" class="btn btn-primary">Logout</button>
