+++ django_backend/game/views.py
@@ -2,6 +2,20 @@ from django.shortcuts import render
 from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
 from django.utils.translation import gettext_lazy as _
+from django.shortcuts import redirect
+from django.utils.translation import activate
+
+# Required for the language
+def set_language(request):
+    if request.method == 'POST':
+        language = request.POST.get('language')
+        if language:
+            request.session[translation.LANGUAGE_SESSION_KEY] = language
+            activate(language)
+
+    next_url = request.POST.get('next') or '/'
+    return redirect(next_url)
+
 
 # Create your views here.
 @login_required
