+++ django_backend/pong_game/settings.py
@@ -49,10 +49,9 @@ INSTALLED_APPS = [
 MIDDLEWARE = [
     'django.middleware.security.SecurityMiddleware',
 	'whitenoise.middleware.WhiteNoiseMiddleware',
-    'django.contrib.sessions.middleware.SessionMiddleware',
-    'django.middleware.common.CommonMiddleware',
+    'django.contrib.sessions.middleware.SessionMiddleware', #for the language support
     'django.middleware.locale.LocaleMiddleware', #for the language support
-    'django.contrib.sessions.middleware.SessionMiddleware', #for language support
+    'django.middleware.common.CommonMiddleware', #for the language support
     'django.middleware.csrf.CsrfViewMiddleware',
     'django.contrib.auth.middleware.AuthenticationMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
